Fall back to individual indexing (slower) if bulk indexing fails
Bulk indexing is way faster than individual indexing, but there are a
few cases where it falls down:

- there's one page-traffic document with an invalid ID, which causes
  the entire batch to be aborted

- the employment_appeal_tribunal_decision, utaac_decision, and
  tax_tribunal_decision types seem to have some large documents, which
  causes a batch to sometimes time out

So when a bulk index fails, try individually indexing every document
in the batch instead.  So we get the speed of bulk indexing, unless it
goes wrong, in which case we then get the robustness of individual
indexing.